feat(accordion): Move accordion, update stories #552
