+++ libs/angular-components/community/components/cards/accordion/accordion-icon/accordion-icon.component.html
