+++ libs/angular-components/community/components/cards/accordion/accordion-item-content/accordion-item-content.component.scss
