+++ libs/angular-components/community/components/cards/accordion/accordion-item/accordion-item.component.html
@@ -2,5 +2,6 @@
   [selected]="selected()"
   class="tedi-accordion-item__card"
   [class.tedi-accordion-item__card--expanded]="opened()"
-  ><ng-content></ng-content
-></tedi-card>
+>
+  <ng-content></ng-content>
+</tedi-card>
