+++ libs/angular-components/community/components/cards/accordion/accordion-item/accordion-item.component.ts
