+++ libs/angular-components/community/components/cards/accordion/accordion/accordion.component.scss
