+++ libs/angular-components/community/components/cards/accordion/index.ts
