+++ libs/angular-components/community/public-api.ts
@@ -1,3 +1,3 @@
 export * from "./components/buttons/button/button.component";
 export * from "./components/cards/card";
-export * from "./components/buttons/accordion";
+export * from "./components/cards/accordion";
